pre-final version

+++ front-end/src/pages/Clients.jsx
@@ -1,48 +1,74 @@
 import React, { useEffect, useState } from "react";
 import "../styles/clients-page.scss";
 import "../styles/loaders.scss"
-import { getClients, newClient, updateClient, deleteClients } from '../services/clients-service';
+import { getClients, deleteClients } from '../services/clients-service';
 import { useDispatch, useSelector } from "react-redux";
 import { updatePagination } from "../store/reducers/clients-reducer";
 import { ClientFieldsMap } from "./client-scheme-fields-map";
 import FormModalWindow from "./components/FormModalWindow";
+import ErrorModalWindow from "../reused/ErrorModalWindow";
 
 const Clients = () => {
 
+    const status = useSelector(state => state.clientsData.status);
     const clients = useSelector(state => state.clientsData.clients);
     const pagination = useSelector(state => state.clientsData.pagination);
     const loading = useSelector(state => state.clientsData.loading);
-    const [totalPageCount, setTotalPageCount] = useState(0);
+    const [totalPageCount, setTotalPageCount] = useState(1);
+    const [dataForModalWindow, setDataForModalWindow] = useState(null);
 
     const [toDelete, setToDelete] = useState([]);
+    const [showErr, toShowErr] = useState(false);
+    const [showEditModal, toShowEditModal] = useState(false);
 
     const dispatch = useDispatch();
     useEffect(() => {
-        const handleDeleteKey = (event) => {
-            if (event.key === 'Delete') {
+        const handleKeys = (event) => {
+            if (event.key === 'Delete' && toDelete.length > 0) {
                 dispatch(deleteClients(toDelete))
                 dispatch(getClients(pagination))
+                setToDelete([]);
+            } else if (event.key === 'Insert') {
+                setDataForModalWindow({ title: 'Новый клиент', defaultData: {} })
+                toShowEditModal(true);
             }
         };
-        window.addEventListener('keydown', handleDeleteKey);
+        window.addEventListener('keyup', handleKeys);
         dispatch(getClients(pagination));
         return () => {
-            window.removeEventListener('keydown', handleDeleteKey);
+            window.removeEventListener('keyup', handleKeys);
         };
     }, [])
 
-    const selectHandle = (id) => {
-        const prev = [...toDelete];
-        const ind = prev.indexOf(id);
-        if (ind < 0) {
-            prev.push(id);
-            setToDelete(prev);
-            document.getElementById(id).style = 'background-color: #969696fd;';
-        } else {
-            prev.splice(ind, 1);
-            setToDelete(prev)
-            document.getElementById(id).style = 'background-color: none';
+    useEffect(() => {
+        if (status === 400) {
+            toShowErr(true);
         }
+    }, [status])
+
+    const selectHandle = (e) => {
+        let clientBlock;
+        if (e.target.className === 'client') {
+            clientBlock = e.target;
+        } else {
+            clientBlock = e.target.closest('.client');
+        }
+        const exists = toDelete.indexOf(clientBlock.id);
+        const current = [...toDelete];
+        if (exists === -1) {
+            clientBlock.style = 'background-color: #969696fd;';
+            current.push(clientBlock.id);
+            setToDelete(current);
+        } else {
+            clientBlock.style = 'background-color: none';
+            current.splice(exists, 1);
+            setToDelete(current);
+        }
+    }
+
+    const openEditFormModal = (title, defaultData) => {
+        setDataForModalWindow({ title, defaultData })
+        toShowEditModal(true);
     }
 
     useEffect(() => {
@@ -53,19 +79,28 @@ const Clients = () => {
 
     const nextPage = (e) => {
         if (pagination.page < totalPageCount) {
-
+            dispatch(updatePagination({ ...pagination, page: pagination.page1 + 1 }))
+            dispatch(getClients(pagination));
         }
     }
 
     const prevPage = (e) => {
         if (pagination.page > 1) {
             dispatch(updatePagination({ ...pagination, page: pagination.page1 - 1 }))
+            dispatch(getClients(pagination));
         }
     }
 
     return (
         <div className="clients-page">
-            <FormModalWindow />
+            {showErr
+                ? <ErrorModalWindow show={toShowErr} />
+                : <></>
+            }
+            {showEditModal
+                ? <FormModalWindow show={toShowEditModal} data={dataForModalWindow} />
+                : <></>
+            }
             <div className="header">
                 Список клиентов
             </div>
@@ -75,17 +110,13 @@ const Clients = () => {
                     : <div className="clients-grid-container">
                         <div className="clients">
                             <div className="client-header">
-                                {clients.length > 0
-                                    ? Object.keys(clients[0]).map((key, index) => {
-                                        if (ClientFieldsMap.has(key)) {
-                                            return <div title={ClientFieldsMap.get(key)} id={index} key={index}>{ClientFieldsMap.get(key)}</div>
-                                        }
-                                    })
-                                    : <></>
+                                {Object.keys(ClientFieldsMap).map((key, index) => {
+                                    return <div title={ClientFieldsMap[key]} id={index} key={index}>{ClientFieldsMap[key]}</div>
+                                })
                                 }
                             </div>
                             {clients?.map((client, index) => {
-                                return <div className="client" id={client.id} onClick={() => selectHandle(client.id)} key={client.id}>
+                                return <div className="client" id={client.id} onClick={selectHandle} key={client.id} onDoubleClick={() => openEditFormModal('Редактирование', client)}>
                                     {Object.keys(client).map((key, index) => {
                                         if (key !== 'id') {
                                             if (typeof client[key] === 'boolean') {
